+++ pkg/crd/crd_test.go
@@ -34,7 +34,7 @@ func TestGenerateYAML(t *testing.T) {
 
 	t.Run("generate multiple CRDs yaml", func(t *testing.T) {
 		// given
-		expected := "apiVersion: executor.testkube.io/v1\nkind: Webhook\nmetadata:\n  name: name1\n  namespace: namespace1\n  labels:\n    key1: value1\nspec:\n  events:\n  - start-test\n  uri: http://localhost\n  selector: app=backend\n\n---\napiVersion: executor.testkube.io/v1\nkind: Webhook\nmetadata:\n  name: name2\n  namespace: namespace2\n  labels:\n    key2: value2\nspec:\n  events:\n  - end-test\n  uri: http://localhost\n  selector: app=backend\n"
+		expected := "apiVersion: executor.testkube.io/v1\nkind: Webhook\nmetadata:\n  name: name1\n  namespace: namespace1\n  labels:\n    key1: value1\nspec:\n  events:\n  - start-test\n  uri: http://localhost\n  selector: app=backend\n\n---\napiVersion: executor.testkube.io/v1\nkind: Webhook\nmetadata:\n  name: name2\n  namespace: namespace2\n  labels:\n    key2: value2\nspec:\n  events:\n  - end-test-success\n  uri: http://localhost\n  selector: app=backend\n"
 		webhooks := []testkube.Webhook{
 			{
 				Name:      "name1",
