+++ pkg/event/bus/nats_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package bus
 
 import (
