+++ pkg/event/loader_test.go
@@ -4,35 +4,17 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/kubeshop/testkube/pkg/event/kind/common"
+	"github.com/kubeshop/testkube/pkg/event/kind/dummy"
 	"github.com/stretchr/testify/assert"
 )
-
-type DummyLoader struct {
-	Err error
-}
-
-func (r DummyLoader) Kind() string {
-	return "dummy"
-}
-
-func (r *DummyLoader) Load() (common.Listeners, error) {
-	if r.Err != nil {
-		return nil, r.Err
-	}
-	return common.Listeners{
-		&DummyListener{},
-		&DummyListener{},
-	}, nil
-}
 
 func TestLoader_UpdateListeners(t *testing.T) {
 
 	t.Run("reconcile updates listeners list based on registered reconcilers", func(t *testing.T) {
 		// given reconciler with two registered reconcilers that return two listeners each
 		reconciler := NewLoader()
-		reconciler.Register(&DummyLoader{})
-		reconciler.Register(&DummyLoader{})
+		reconciler.Register(&dummy.DummyLoader{})
+		reconciler.Register(&dummy.DummyLoader{})
 
 		// when
 		listeners := reconciler.Reconcile()
@@ -44,8 +26,8 @@ func TestLoader_UpdateListeners(t *testing.T) {
 	t.Run("reconcile updates listeners list based on registered reconcilers thread safe", func(t *testing.T) {
 		// given reconciler with two registered reconcilers that return two listeners each
 		reconciler := NewLoader()
-		reconciler.Register(&DummyLoader{})
-		reconciler.Register(&DummyLoader{})
+		reconciler.Register(&dummy.DummyLoader{})
+		reconciler.Register(&dummy.DummyLoader{})
 
 		// when
 		listeners := reconciler.Reconcile()
@@ -57,8 +39,8 @@ func TestLoader_UpdateListeners(t *testing.T) {
 	t.Run("failed loaders are omited", func(t *testing.T) {
 		// given reconciler with two registered reconcilers that return two listeners each
 		reconciler := NewLoader()
-		reconciler.Register(&DummyLoader{Err: fmt.Errorf("loader error")})
-		reconciler.Register(&DummyLoader{})
+		reconciler.Register(&dummy.DummyLoader{Err: fmt.Errorf("loader error")})
+		reconciler.Register(&dummy.DummyLoader{})
 
 		// when
 		listeners := reconciler.Reconcile()
